Rework viewport component
