--- ui/mat/theme.go
@@ -1,38 +1,0 @@
-package mat
-
-import "github.com/mokiat/lacking/ui"
-
-var (
-	PrimaryColor   ui.Color = ui.RGB(0x40, 0x89, 0xBD)
-	OnPrimaryColor ui.Color = ui.RGB(0x00, 0x00, 0x00)
-
-	PrimaryLightColor   ui.Color = ui.RGB(0x7C, 0xAE, 0xEB)
-	OnPrimaryLightColor ui.Color = ui.RGB(0x00, 0x00, 0x00)
-
-	PrimaryDarkColor   ui.Color = ui.RGB(0x32, 0x59, 0x73)
-	OnPrimaryDarkColor ui.Color = ui.RGB(0xFF, 0xFF, 0xFF)
-
-	SecondaryColor   ui.Color = ui.RGB(0x8B, 0xC3, 0x4A)
-	OnSecondaryColor ui.Color = ui.RGB(0x00, 0x00, 0x00)
-
-	SecondaryLightColor   ui.Color = ui.RGB(0xA9, 0xC2, 0x8A)
-	OnSecondaryLightColor ui.Color = ui.RGB(0x00, 0x00, 0x00)
-
-	SecondaryDarkColor   ui.Color = ui.RGB(0x4B, 0x6B, 0x3A)
-	OnSecondaryDarkColor ui.Color = ui.RGB(0xFF, 0xFF, 0xFF)
-
-	BackgroundColor   ui.Color = ui.RGB(0x3B, 0x3B, 0x3B)
-	OnBackgroundColor ui.Color = ui.RGB(0xFF, 0xFF, 0xFF)
-
-	SurfaceColor   ui.Color = ui.RGB(0xFF, 0xFF, 0xFF)
-	OnSurfaceColor ui.Color = ui.RGB(0x00, 0x00, 0x00)
-
-	ErrorColor   ui.Color = ui.RGB(0xB3, 0x1E, 0x00)
-	OnErrorColor ui.Color = ui.RGB(0xFF, 0xFF, 0xFF)
-
-	OutlineColor ui.Color = ui.RGB(0xB0, 0xB0, 0xB0)
-
-	HoverOverlayColor ui.Color = ui.RGBA(0x00, 0x00, 0x00, 0x40)
-	PressOverlayColor ui.Color = ui.RGBA(0x00, 0x00, 0x00, 0x80)
-	ModalOverlayColor ui.Color = ui.RGBA(0x00, 0x00, 0x00, 0xA0)
-)
